Moved web/middleware to web/shared/middleware
