+++ core/server/web/api/v2/admin/app.js
@@ -9,13 +9,13 @@ const routes = require('./routes');
 // Include the middleware
 
 // API specific
-const versionMatch = require('../../../middleware/api/version-match'); // global
+const versionMatch = require('../../../shared/middleware/api/version-match'); // global
 
 // Shared
 const bodyParser = require('body-parser'); // global, shared
-const cacheControl = require('../../../middleware/cache-control'); // global, shared
-const maintenance = require('../../../middleware/maintenance'); // global, shared
-const errorHandler = require('../../../middleware/error-handler'); // global, shared
+const cacheControl = require('../../../shared/middleware/cache-control'); // global, shared
+const maintenance = require('../../../shared/middleware/maintenance'); // global, shared
+const errorHandler = require('../../../shared/middleware/error-handler'); // global, shared
 
 module.exports = function setupApiApp() {
     debug('Admin API v2 setup start');
