+++ core/server/web/shared/middleware/error-handler.js
@@ -1,8 +1,8 @@
 const hbs = require('express-hbs');
 const debug = require('ghost-ignition').debug('error-handler');
-const config = require('../../config');
-const common = require('../../lib/common');
-const helpers = require('../../services/routing/helpers');
+const config = require('../../../config');
+const common = require('../../../lib/common');
+const helpers = require('../../../services/routing/helpers');
 
 const escapeExpression = hbs.Utils.escapeExpression;
 const _private = {};
@@ -14,7 +14,7 @@ const errorHandler = {};
  */
 _private.createHbsEngine = () => {
     const engine = hbs.create();
-    engine.registerHelper('asset', require('../../helpers/asset'));
+    engine.registerHelper('asset', require('../../../helpers/asset'));
 
     return engine.express4();
 };
