+++ core/server/web/shared/middleware/frontend-client.js
@@ -1,6 +1,6 @@
-const api = require('../../api');
-const labs = require('../../services/labs');
-const common = require('../../lib/common');
+const api = require('../../../api');
+const labs = require('../../../services/labs');
+const common = require('../../../lib/common');
 
 module.exports = function getFrontendClient(req, res, next) {
     if (labs.isSet('publicAPI') !== true) {
