+++ core/server/web/shared/middleware/image/index.js
