+++ core/server/web/shared/middleware/index.js
