+++ core/server/web/shared/middleware/labs.js
@@ -1,5 +1,5 @@
-const labsUtil = require('../../services/labs');
-const common = require('../../lib/common');
+const labsUtil = require('../../../services/labs');
+const common = require('../../../lib/common');
 
 const labs = {
     subscribers(req, res, next) {
