+++ core/server/web/shared/middleware/pretty-urls.js
@@ -7,7 +7,7 @@
 // @TODO optimise this to reduce the number of redirects required to get to a pretty URL
 // @TODO move this to being used by routers?
 const slashes = require('connect-slashes');
-const config = require('../../config');
+const config = require('../../../config');
 
 module.exports = [
     slashes(true, {
