+++ core/server/web/shared/middleware/static-theme.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const path = require('path');
-const config = require('../../config');
-const constants = require('../../lib/constants');
-const themeUtils = require('../../services/themes');
+const config = require('../../../config');
+const constants = require('../../../lib/constants');
+const themeUtils = require('../../../services/themes');
 
 function isBlackListedFileType(file) {
     const blackListedFileTypes = ['.hbs', '.md', '.json'];
