+++ core/server/web/shared/middleware/url-redirects.js
@@ -1,7 +1,7 @@
 const url = require('url');
 const path = require('path');
 const debug = require('ghost-ignition').debug('url-redirects');
-const urlService = require('../../services/url');
+const urlService = require('../../../services/url');
 
 const _private = {};
