+++ core/test/unit/web/middleware/serve-public-file_spec.js
@@ -1,7 +1,7 @@
 var should = require('should'),
     sinon = require('sinon'),
     fs = require('fs-extra'),
-    servePublicFile = require('../../../../server/web/middleware/serve-public-file'),
+    servePublicFile = require('../../../../server/web/shared/middleware/serve-public-file'),
 
     sandbox = sinon.sandbox.create();
