+++ core/test/unit/web/middleware/static-theme_spec.js
@@ -3,7 +3,7 @@ var should = require('should'),
 
     express = require('express'),
     themeUtils = require('../../../../server/services/themes'),
-    staticTheme = require('../../../../server/web/middleware/static-theme'),
+    staticTheme = require('../../../../server/web/shared/middleware/static-theme'),
 
     sandbox = sinon.sandbox.create();
