+++ core/test/unit/web/middleware/url-redirects_spec.js
@@ -2,7 +2,7 @@ var should = require('should'),
     sinon = require('sinon'),
     rewire = require('rewire'),
     configUtils = require('../../../utils/configUtils'),
-    urlRedirects = rewire('../../../../server/web/middleware/url-redirects'),
+    urlRedirects = rewire('../../../../server/web/shared/middleware/url-redirects'),
     {adminRedirect} = urlRedirects,
     getAdminRedirectUrl = urlRedirects.__get__('_private.getAdminRedirectUrl'),
     getBlogRedirectUrl = urlRedirects.__get__('_private.getBlogRedirectUrl'),
