+++ core/test/utils/index.js
@@ -23,7 +23,7 @@ var Promise = require('bluebird'),
     settingsService = require('../../server/services/settings'),
     settingsCache = require('../../server/services/settings/cache'),
     imageLib = require('../../server/lib/image'),
-    customRedirectsMiddleware = require('../../server/web/middleware/custom-redirects'),
+    customRedirectsMiddleware = require('../../server/web/shared/middleware/custom-redirects'),
     permissions = require('../../server/services/permissions'),
     sequence = require('../../server/lib/promise/sequence'),
     themes = require('../../server/services/themes'),
